backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9

python-dotenv==1.0.1
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2

httpx==0.27.2

ultralytics==8.3.0
opencv-python-headless==4.10.0.84
